requirements.txt
python-dotenv-0.19.2
pyTelegramBotAPI-4.4.0
